fix: adds minor structural and renaming changes
