+++ tests/operations/test_snowflake.py
@@ -1,13 +1,20 @@
+import base64
 from unittest.mock import MagicMock, patch
 
 import pytest
 import snowflake.connector
+from pydantic import SecretStr
 
+from knowledge_graph.cloud import AwsEnv
 from knowledge_graph.operations.snowflake import (
+    SNOWFLAKE_ACCOUNT_SSM,
     SNOWFLAKE_DATABASE,
+    SNOWFLAKE_PRIVATE_KEY_SSM,
     SNOWFLAKE_SCHEMA,
+    SNOWFLAKE_USER_SSM,
     SNOWFLAKE_WAREHOUSE,
     connect_to_snowflake,
+    get_snowflake_credentials,
 )
 
 
@@ -66,3 +73,36 @@ def test_connect_reraises_local_connection_error():
     ):
         with pytest.raises(snowflake.connector.errors.Error):
             connect_to_snowflake()
+
+
+def test_get_snowflake_credentials_resolves_and_decodes():
+    ssm_values = {
+        SNOWFLAKE_ACCOUNT_SSM: "test_account",
+        SNOWFLAKE_USER_SSM: "svc_user",
+        # Private key is stored base64-encoded in SSM and decoded on the way out.
+        SNOWFLAKE_PRIVATE_KEY_SSM: base64.b64encode(b"PEM_KEY_DATA").decode("utf-8"),
+    }
+
+    with patch(
+        "knowledge_graph.operations.snowflake.get_aws_ssm_param",
+        side_effect=lambda name, aws_env: ssm_values[name],
+    ):
+        account, user, private_key = get_snowflake_credentials(AwsEnv.production)
+
+    assert account == "test_account"
+    assert user == "svc_user"
+    assert isinstance(private_key, SecretStr)
+    assert private_key.get_secret_value() == "PEM_KEY_DATA"
+
+
+def test_get_snowflake_credentials_passes_aws_env_through():
+    with patch(
+        "knowledge_graph.operations.snowflake.get_aws_ssm_param",
+        return_value=base64.b64encode(b"x").decode("utf-8"),
+    ) as mock_ssm:
+        get_snowflake_credentials(AwsEnv.staging)
+
+    assert mock_ssm.call_count == 3
+    assert all(
+        call.kwargs["aws_env"] == AwsEnv.staging for call in mock_ssm.call_args_list
+    )
